requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
Django==3.2.8
django-livereload-server==0.3.2
djangorestframework==3.12.4
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
